Complete Controller with * view, edit, create, destroy * fix permission check und

+++ src/Http/Controllers/SeatGroupsController.php
@@ -3,12 +3,20 @@
 namespace Herpaderpaldent\Seat\SeatGroups\Http\Controllers;
 
 use Herpaderpaldent\Seat\SeatGroups\Models\Seatgroup;
+use Herpaderpaldent\Seat\SeatGroups\Models\Seatgroupmanager;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Seat\Eveapi\Models\Character\CharacterInfo;
+use Seat\Services\Repositories\Character\Character;
+use Seat\Services\Repositories\Configuration\UserRespository;
+use Seat\Services\Repositories\Corporation\Corporation;
+use Seat\Web\Acl\AccessManager;
 use Seat\Web\Http\Controllers\Controller;
 use Seat\Web\Models\Acl\Role;
 
 class SeatGroupsController extends Controller
 {
+    use AccessManager, UserRespository, Character, Corporation;
     /**
      * Display a listing of the resource.
      *
@@ -52,9 +60,10 @@ class SeatGroupsController extends Controller
             'role_id' => 'numeric'
         ]);
 
-        Seatgroup::create($seatgroup);
+        $group=Seatgroup::create($seatgroup);
         //ToDo: if logic implementation for failed validation + forward to view
-        return back()->with('success', 'SeAT-Group has been added');
+        return redirect()->route('seatgroups.edit', $group->id)
+            ->with('success', 'SeAT-Group has been added');
     }
 
     /**
@@ -76,18 +85,22 @@ class SeatGroupsController extends Controller
      */
     public function edit($id)
     {
-        //
-        /*$Roles=Role::pluck('title','id');
 
-        if(Seatgroup::find($id)->isManager(auth()->user(),$id) || auth()->user()->has('Superuser')){
-            return view('seatgroups::edit')
-                ->with('group',SeatGroup::find($id))
-                ->with('roles',$Roles);
-        }*/
+        $all_users = $this->getAllUsers()
+            ->pluck('name')
+            ->toArray();
+
+        //$all_characters = $this->getAllCharacters();
+
+
+        //$all_corporations = $this->getAllCorporations();
+        $all_corporations = $this->getAllCharacters(); // TODO: Remove this line
+
+
         $seatgroup = Seatgroup::find($id);
         $Roles=Role::pluck('title','id');
 
-        return view('seatgroups::edit', compact('seatgroup','id'))
+        return view('seatgroups::edit', compact('seatgroup','id','all_corporations'))
             ->with('roles',$Roles);
 
     }
@@ -126,9 +139,13 @@ class SeatGroupsController extends Controller
      */
     public function destroy($id)
     {
-        $seatgroup = Seatgroup::find($id);
-        $seatgroup->delete();
+        if(Auth::user()->hasRole('Superuser')){
+            $seatgroupmanager = Seatgroupmanager::with('group_id','=',$id);
+            $seatgroupmanager->delete();
+            $seatgroup = Seatgroup::find($id);
+            $seatgroup->delete();
         return redirect()->route('seatgroups.index')
             ->with('success', 'SeAT-Group has been deleted');
+        } else return redirect()->back()->with('error', 'illegal delete request. You must be superuser');
     }
 }
